+++ include/NOTGate.h
@@ -4,7 +4,7 @@
 class NOTGate: public Gate{
 private:
 public:
-	NOTGate(int propagationTime) : Gate(propagationTime) {}; //dodac Gate(int n) gdzie n to liczba wejsc bramki
+	NOTGate(double propagationTime) : Gate(propagationTime) {}; //dodac Gate(int n) gdzie n to liczba wejsc bramki
 	void propagateOutput() override;
 	std::string returnType() override;
 	int getNumberOfInputs() override;
