+++ include/ORGate.h
@@ -3,7 +3,7 @@
 class ORGate : public Gate {
 private:
 public:
-	ORGate(int propagationTime) : Gate(propagationTime) {};
+	ORGate(double propagationTime) : Gate(propagationTime) {};
 	void propagateOutput() override;
 	std::string returnType() override;
 	int getNumberOfInputs() override;
