+++ include/XORGate.h
@@ -3,7 +3,7 @@
 class XORGate : public Gate {
 private:
 public:
-	XORGate(int propagationTime) : Gate(propagationTime) {};
+	XORGate(double propagationTime) : Gate(propagationTime) {};
 	void propagateOutput() override;
 	std::string returnType() override;
 	int getNumberOfInputs() override;
